Fix provisioning "no submit uri" test failure
The error message in this case appears to have changed not to include
the DNS resolver address, resulting in the expected error failure.

To get around that and make the test more robust, check just for the "no
such host" fragment inside the error message, rather than attempting to
match the entire message.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>